itns to float

+++ pong-master/Pong/Pong/game.cpp
@@ -82,7 +82,7 @@ void game::Render()
 
 void game::RenderBall() 
 {
-	Ball = { ballX, ballY, 10, 10 };
+	Ball = { (int)ballX, (int)ballY, 10, 10 };
 	SDL_SetRenderDrawColor(renderer, 255, 255, 255, 0);
 	SDL_RenderFillRect(renderer, &Ball);
 }
